Extract world hop service
